Add quick emoji button next to text input
This allows for a user selectable button to be displayed next to the
SMS text input area to allow quick access to the emoji icons.

Seeing as how the current process to add an emoji requires going through
the menu and then selecting the option, this allows for frequent users
of emojis a much simpler and time saving method.

This modification is based heavily off the AOKP source
introduced by Roman Birg.

New drawable: Jacob Müller

Conflicts:

	res/values/strings.xml
	res/xml/preferences.xml
	src/com/android/mms/ui/MessagingPreferenceActivity.java

Change-Id: I6fab11f155c5c112ed39785179093cecaa5500f6